@@ -6,23 +6,23 @@
   "scripts": {
     "build": "yarn clean && tsc",
     "clean": "rimraf dist",
-    "start": "node dist/app.js",
-    "test": "jest --verbose --detectOpenHandles --forceExit"
+    "start": "node dist/relay.js",
+    "starts": "node dist/relay.js https",
+    "sniff": "node dist/sniffer.js",
+    "mock-endpoint": "node dist/__mocks__/mockEndpoint.js",
+    "test": "jest --verbose --detectOpenHandles --runInBand --forceExit"
   },
   "private": true,
   "dependencies": {
-    "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "http-proxy-middleware": "^1.0.4",
     "json5": "^2.1.3",
-    "loadash": "^1.0.0",
+    "lodash": "^4.17.15",
     "mockyeah": "^0.24.1",
-    "node-redis-retry-strategy": "^1.0.10",
     "query-string": "^6.12.1",
-    "randomstring": "^1.1.5",
-    "redis": "^3.0.2",
-    "util": "^0.12.3",
+    "ioredis": "^4.17.3",
+    "stoppable": "^1.1.0",
     "winston": "^3.2.1"
   },
   "dependenciesComments": {
@@ -30,15 +30,17 @@
   },
   "devDependencies": {
     "@types/json5": "0.0.30",
+    "@types/lodash": "^4.14.152",
     "@types/randomstring": "^1.1.6",
     "@types/redis": "^2.8.21",
+    "@types/ioredis": "^4.16.4",
+    "@types/stoppable": "^1.1.0",
     "jest": "^24.9.0",
     "mockyeah": "^0.24.1",
     "rimraf": "^3.0.2",
     "run-container": "^1.2.1",
     "supertest": "^4.0.2",
     "ts-jest": "^24.0.2",
-    "ts-node": "^8.3.0",
-    "typescript": "^3.7.2"
+    "typescript": "^3.8.3"
   }
 }
