@@ -1,8 +1,6 @@
 require('./env');
-import http from 'http';
 import util from 'util';
 import { createGateway, getLogger } from '@fabric-es/gateway-lib';
-import stoppable from 'stoppable';
 
 const PORT = (process.env.GATEWAY_PORT || 4001) as number;
 const authenticationCheck = process.env.AUTHORIZATION_SERVER_URI;
@@ -11,7 +9,7 @@ const logger = getLogger('[gw-org1] app.js');
 (async () => {
   logger.info('♨️♨️  Starting [gw-org1] gateway');
 
-  const server = await createGateway({
+  const { gateway, shutdown } = await createGateway({
     serviceList: [
       {
         name: 'user',
@@ -44,18 +42,6 @@ const logger = getLogger('[gw-org1] app.js');
     debug: false
   });
 
-  const stoppableServer = stoppable(http.createServer(server));
-
-  const shutdown = () => {
-    stoppableServer.close(err => {
-      if (err) {
-        logger.error(util.format('An error occurred while closing the server: %j', err));
-        process.exitCode = 1;
-      } else logger.info('server closes');
-    });
-    process.exit();
-  };
-
   process.on('SIGINT', () => {
     shutdown();
   });
@@ -69,7 +55,7 @@ const logger = getLogger('[gw-org1] app.js');
     logger.error(err.stack);
   });
 
-  stoppableServer.listen(PORT, () => {
+  gateway.listen(PORT, () => {
     logger.info(`🚀 gateway ready at http://${process.env.GATEWAY_HOST}:${PORT}/graphql`);
     if (process.env.NODE_ENV === 'production') process.send('ready');
   });
