@@ -1,27 +1,37 @@
-import _ from 'lodash';
-import { ReqRes } from './reqres';
+import { Redis } from 'ioredis';
+import isEmpty from 'lodash/isEmpty';
 import { getLogger } from './getLogger';
+import { ReqRes } from './reqres';
 
 const logger = getLogger('[relay] processMsg.js');
 
-export const processMsg = async ({
-  message, 
-  client, 
-  topic
+export const processMsg = ({
+  message,
+  client,
+  topic,
+  ttl,
 }: {
-  message: ReqRes; 
-  client: any; 
-  topic: string
+  message: ReqRes;
+  client: Redis;
+  topic: string;
+  ttl?: number;
 }) => {
+  return new Promise<number>((resolve, reject) => {
+    if (isEmpty(message))
+      reject(new Error('Message missing'));
+    else if (isEmpty(client))
+      reject(new Error('Client missing'));
+    else if (isEmpty(topic))
+      reject(new Error('Topic missing'));
+    else {
+      const messageStr = JSON.stringify(message);
+      const timestamp = Date.now();
+      const offset = ttl ? ttl : 86400000; // 1 day == 24x60x60x1000 milliseconds
 
-  if (_.isEmpty(message)) throw new Error('Missing message.');
-  if (_.isEmpty(client)) throw new Error('Missing client');
-  if (_.isEmpty(topic)) throw new Error('Missing topic.');
+      client.zremrangebyscore(topic, '-inf', (timestamp - offset));
+      client.zadd(topic, timestamp, messageStr);
 
-  client.on('error', (err) => {
-    throw new Error(`Publish client error: ${err}`);
+      client.publish(topic, messageStr).then(value => resolve(value));
+    }
   });
-
-  const messageStr = JSON.stringify(message);
-  await client.publish(topic, messageStr, () => { logger.info(`Published message: ${messageStr}`) });
 };
